+++ Changelog.md
@@ -3,7 +3,7 @@
 ## Version (NEXT)
 * use a single state atom in simulator
 * added perfect power factor to math
-
+* generalized quantum phase estimation and integrated in quantum-period-finding
 
 ## Version 0.1.0
 * implemented quantom states, quantum gates, gate registry and quantum cirquits